Update packaging of API endpoint instrumentation module com.newrelic.agent.security.instrumentation.*
